Attendance System using Face recognisation
This project is a Face Recognition Attendance System built using Python. It automates the process of marking attendance by recognizing faces from a webcam feed in real-time. The system captures facial images of students, trains a model to recognize them, and marks their attendance when they appear in front of the camera.